+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.2] - 2022-12-25
+
 ### Added
 
 - Repeat lookups until a sureness of at least 80.0% is reached
@@ -42,5 +44,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the MIT license
 - Created a CHANGELOG
 
-[Unreleased]: https://github.com/e-zim/emporia-time/compare/v0.0.1...HEAD
+[Unreleased]: https://github.com/e-zim/emporia-time/compare/v0.0.2...HEAD
+[0.0.2]: https://github.com/e-zim/emporia-time/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/e-zim/emporia-time/releases/tag/v0.0.1